assignment 3

+++ routes/dishRouter.js
@@ -2,6 +2,7 @@ const express=require("express")
 const bodyparser=require("body-parser")
 const mongoose=require("mongoose")
 const Dishes=require("../models/dishes")
+const authenticate = require("../authenticate")
 
 
 const dishRouter=express.Router()
@@ -10,6 +11,7 @@ dishRouter.use(bodyparser.json())
 dishRouter.route("/")
 .get((req,res,next)=>{
     Dishes.find({})
+    .populate("comments.auther")
     .then((dishes)=>{
         res.statusCode=200
         res.setHeader("Content-Type","application/json")
@@ -17,28 +19,42 @@ dishRouter.route("/")
     },(err)=>console.log(err))
     .catch((err)=>next(err))
 })
-.post((req,res,next)=>{
-    Dishes.create(req.body)
-    .then((dish)=>{
-        console.log("Dish created",dish)
-        res.statusCode=200
-        res.setHeader("Content-Type","application/json")
-        res.json(dish)
-    },(err)=>console.log(err))
-    .catch((err)=>next(err))
+.post(authenticate.verifyUser, (req,res,next)=>{
+    if(authenticate.verifyAdmin({user: req.user})){
+        Dishes.create(req.body)
+        .then((dish)=>{
+            res.statusCode=200
+            res.setHeader("Content-Type","application/json")
+            res.json(dish)
+        },(err)=>console.log(err))
+        .catch((err)=>next(err))
+    }
+    else{
+        res.statusCode=403
+        res.setHeader("Content-Type","plain/text")
+        res.json("Not authorized!")
+    }
+    
 })
-.put((req,res,next)=>{
+.put(authenticate.verifyUser , (req,res,next)=>{
     res.statusCode=403
     res.end("PUT command is not supported for /dishes !")
 })
-.delete((req,res,next)=>{
-    Dishes.remove({})
-    .then((resp)=>{
-        res.statusCode=200
-        res.setHeader("Content-Type","application/json")
-        res.json(resp)
-    },(err)=>console.log(err))
-    .catch((err)=>next(err))
+.delete(authenticate.verifyUser ,(req,res,next)=>{
+    if(authenticate.verifyAdmin({user: req.user})){
+        Dishes.remove({})
+        .then((resp)=>{
+            res.statusCode=200
+            res.setHeader("Content-Type","application/json")
+            res.json(resp)
+        },(err)=>console.log(err))
+        .catch((err)=>next(err))
+}
+else{
+    res.statusCode=403
+    res.setHeader("Content-Type","plain/text")
+    res.json("Not authorized!")
+}
 })
 
 
@@ -48,6 +64,7 @@ dishRouter.route("/")
 dishRouter.route("/:dishId")
 .get((req,res,next)=>{
     Dishes.findById(req.params.dishId)
+    .populate("comments.auther")
     .then((dish)=>{
         res.statusCode=200
         res.setHeader("Content-Type","application/json")
@@ -55,30 +72,44 @@ dishRouter.route("/:dishId")
     },(err)=>console.log(err))
     .catch((err)=>next(err))
 })
-.post((req,res,next)=>{
+.post(authenticate.verifyUser, (req,res,next)=>{
     res.statusCode=403
     res.end("POST command is not supported for /dishes/"+req.params.dishId)
 })
-.put((req,res,next)=>{
-    Dishes.findByIdAndUpdate(req.params.dishId,{
-        $set:req.body
-    },
-    {new:true})
-    .then((dish)=>{
-        res.statusCode=200
-        res.setHeader("Content-Type","application/json")
-        res.json(dish)
-    },(err)=>console.log(err))
-    .catch((err)=>next(err))
+.put(authenticate.verifyUser, (req,res,next)=>{
+    if(authenticate.verifyAdmin({user: req.user})){
+        Dishes.findByIdAndUpdate(req.params.dishId,{
+            $set:req.body
+        },
+        {new:true})
+        .then((dish)=>{
+            res.statusCode=200
+            res.setHeader("Content-Type","application/json")
+            res.json(dish)
+        },(err)=>console.log(err))
+        .catch((err)=>next(err))
+    }
+    else{
+        res.statusCode=403
+        res.setHeader("Content-Type","plain/text")
+        res.json("Not authorized!")
+    }
 })
-.delete((req,res,next)=>{
-    Dishes.findByIdAndRemove(req.params.dishId)
-    .then((resp)=>{
-        res.statusCode=200
-        res.setHeader("Content-Type","application/json")
-        res.json(resp)
-    },(err)=>console.log(err))
-    .catch((err)=>next(err))
+.delete(authenticate.verifyUser, (req,res,next)=>{
+    if(authenticate.verifyAdmin({user: req.user})){
+        Dishes.findByIdAndRemove(req.params.dishId)
+        .then((resp)=>{
+            res.statusCode=200
+            res.setHeader("Content-Type","application/json")
+            res.json(resp)
+        },(err)=>console.log(err))
+        .catch((err)=>next(err))
+    }
+    else{
+        res.statusCode=403
+        res.setHeader("Content-Type","plain/text")
+        res.json("Not authorized!")
+    }
 })
 
 
@@ -88,6 +119,7 @@ dishRouter.route("/:dishId")
 dishRouter.route("/:dishId/comments")
 .get((req,res,next)=>{
     Dishes.findById(req.params.dishId)
+    .populate("comments.auther")
     .then((dish)=>{
         if(dish != null){
             res.statusCode=200
@@ -102,16 +134,21 @@ dishRouter.route("/:dishId/comments")
     },(err)=>console.log(err))
     .catch((err)=>next(err))
 })
-.post((req,res,next)=>{
+.post(authenticate.verifyUser, (req,res,next)=>{
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
         if(dish != null){
+            req.body.auther = req.user._id
             dish.comments.push(req.body)
             dish.save()
             .then((dish)=>{
-                res.statusCode=200
-                res.setHeader("Content-Type","application/json")
-                res.json(dish.comments)
+                Dishes.findById(dish._id)
+                    .populate("comments.auther")
+                    .then((dish)=>{
+                        res.statusCode=200
+                        res.setHeader("Content-Type","application/json")
+                        res.json(dish.comments)
+                    })
             })
         }
         else{
@@ -122,30 +159,38 @@ dishRouter.route("/:dishId/comments")
     },(err)=>console.log(err))
     .catch((err)=>next(err))
 })
-.put((req,res,next)=>{
+.put(authenticate.verifyUser, (req,res,next)=>{
     res.statusCode=403
     res.end("POST command is not supported for /dishes/"+req.params.dishId+"comments !")
 })
-.delete((req,res,next)=>{
-    Dishes.findById(req.params.dishId)
-    .then((dish)=>{
-        if(dish != null){
-            for(let i=(dish.comments.length)-1 ; i>=0 ;i--)
-                dish.comments.id(dish.comments[i]._id).remove()
-            dish.save()
-            .then((dish)=>{
-                res.statusCode=200
-                res.setHeader("Content-Type","application/json")
-                res.json(dish.comments)
-            })
-        }
-        else{
-            const err=new Error("Dish "+req.params.dishId+" has not found!")
-            err.status=404
-            return next(err)
-        }
-    },(err)=>console.log(err))
-    .catch((err)=>next(err))
+.delete(authenticate.verifyUser, (req,res,next)=>{
+    if(authenticate.verifyAdmin({user: req.user})){
+        Dishes.findById(req.params.dishId)
+        .then((dish)=>{
+            if(dish != null){
+                for(let i=(dish.comments.length)-1 ; i>=0 ;i--)
+                    dish.comments.id(dish.comments[i]._id).remove()
+                dish.save()
+                .then((dish)=>{
+                    res.statusCode=200
+                    res.setHeader("Content-Type","application/json")
+                    res.json(dish.comments)
+                })
+            }
+            else{
+                const err=new Error("Dish "+req.params.dishId+" has not found!")
+                err.status=404
+                return next(err)
+            }
+        },(err)=>console.log(err))
+        .catch((err)=>next(err))
+    }
+    else{
+        res.statusCode=403
+        res.setHeader("Content-Type","plain/text")
+        res.json("Not authorized!")
+    }
+    
 })
 
 
@@ -154,6 +199,7 @@ dishRouter.route("/:dishId/comments")
 dishRouter.route("/:dishId/comments/:commentId")
 .get((req,res,next)=>{
     Dishes.findById(req.params.dishId)
+    .populate("comments.auther")
     .then((dish)=>{
         if(dish != null && dish.comments.id(req.params.commentId)!=null){
             res.statusCode=200
@@ -173,64 +219,86 @@ dishRouter.route("/:dishId/comments/:commentId")
     },(err)=>console.log(err))
     .catch((err)=>next(err))
 })
-.post((req,res,next)=>{
+.post(authenticate.verifyUser, (req,res,next)=>{
     res.statusCode=403
     res.end("POST command is not supported for /dishes/"+req.params.dishId+"/comments/"+req.params.commentId)
 })
-.put((req,res,next)=>{
+.put(authenticate.verifyUser, (req,res,next)=>{
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
-        if(dish != null && dish.comments.id(req.params.commentId) != null){
-            if(req.body.rating) 
-            dish.comments.id(req.params.commentId).rating=req.body.rating
-
-            if(req.body.comment) 
-            dish.comments.id(req.params.commentId).comment=req.body.comment
-
-            dish.save()
-            .then(()=>{
-                res.statusCode=200
-                res.setHeader("Content-Type","application/json")
-                res.json(dish)
-            },err => next(err))
+            if(authenticate.verifyEligibility(req.user._id,dish.comments.id(req.params.commentId).auther._id)){
+            if(dish != null && dish.comments.id(req.params.commentId) != null){
+                if(req.body.rating) 
+                dish.comments.id(req.params.commentId).rating=req.body.rating
+    
+                if(req.body.comment) 
+                dish.comments.id(req.params.commentId).comment=req.body.comment
+    
+                dish.save()
+                .then((dish)=>{
+                    Dishes.findById(dish._id)
+                        .populate("comments.auther")
+                        .then((dish)=>{
+                            res.statusCode=200
+                            res.setHeader("Content-Type","application/json")
+                            res.json(dish)
+                        })
+                },err => next(err))
+            }
+            else if(dish == null){
+                const err=new Error("Dish "+req.params.dishId+" has not found!")
+                err.status=404
+                return next(err)
+            }
+            else if(dish.comments.id(req.params.commentId) == null){
+                const err=new Error("Comment "+req.params.commentId+" has not found!")
+                err.status=404
+                return next(err)
+            }
         }
-        else if(dish == null){
-            const err=new Error("Dish "+req.params.dishId+" has not found!")
-            err.status=404
-            return next(err)
+        else{
+            res.statusCode=403
+            res.setHeader("Content-Type","plain/text")
+            res.json("Not authorized!") 
         }
-        else if(dish.comments.id(req.params.commentId) == null){
-            const err=new Error("Comment "+req.params.commentId+" has not found!")
-            err.status=404
-            return next(err)
-        }
-    }, err => console.log(err))
-    .catch( err => next(err))
+        }, err => console.log(err))
+        .catch( err => next(err)) 
 })
-.delete((req,res,next)=>{
+.delete(authenticate.verifyUser, (req,res,next)=>{
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
-        if(dish != null && dish.comments.id(req.params.commentId) != null){
-            dish.comments.id(req.params.commentId).remove()
-            dish.save()
-            .then(()=>{
-                res.statusCode=200
-                res.setHeader("Content-Type","application/json")
-                res.json(dish.comments)
-            })
-        }
-        else if(dish == null){
-            const err=new Error("Dish "+req.params.dishId+" has not found!")
-            err.status=404
-            return next(err)
-        }
-        else if(dish.comments.id(req.params.commentId) == null){
-            const err=new Error("Comment "+req.params.commentId+" has not found!")
-            err.status=404
-            return next(err)
-        }
-    }, err => console.log(err))
-    .catch( err => next(err))
+            if(authenticate.verifyEligibility(req.user._id,dish.comments.id(req.params.commentId).auther._id)){
+                if(dish != null && dish.comments.id(req.params.commentId) != null){
+                    dish.comments.id(req.params.commentId).remove()
+                    dish.save()
+                    .then((dish)=>{
+                        Dishes.findById(dish._id)
+                            .populate("comments.auther")
+                            .then((dish)=>{
+                                res.statusCode=200
+                                res.setHeader("Content-Type","application/json")
+                                res.json(dish)
+                            })
+                    })
+                }
+                else if(dish == null){
+                    const err=new Error("Dish "+req.params.dishId+" has not found!")
+                    err.status=404
+                    return next(err)
+                }
+                else if(dish.comments.id(req.params.commentId) == null){
+                    const err=new Error("Comment "+req.params.commentId+" has not found!")
+                    err.status=404
+                    return next(err)
+                }
+            }
+            else{
+                res.statusCode=403
+                res.setHeader("Content-Type","plain/text")
+                res.json("Not authorized!") 
+            }
+        }, err => console.log(err))
+        .catch( err => next(err))
 })
